added wait for ready and someother housekeeping
